refactor: move validation logics to widget-validation-helper
